backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
psycopg2-binary==2.9.10
pillow==10.4.0
boto3==1.34.23
django-storages==1.14.2
requests==2.31.0
celery==5.3.4
redis==5.0.1
paystack==1.5.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==21.0.0
drf-spectacular==0.26.5
django-ratelimit==4.1.0
django-filter==23.5
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.39.1
markdown==3.5.1
django-extensions==3.2.3
daphne==4.0.0
